backend/pyproject.toml
[tool.poetry]
name = "ChannelExplorer"
version = "0.0.2"
description = ""
# authors = ["Rahat Zaman <rahatzamancse@gmail.com>"]
authors = ["anonymous <anonymous@gmail.com>"]
# readme = "../README.md"
homepage = "https://github.com/anonymous/ChannelExplorer"
repository = "https://github.com/anonymous/ChannelExplorer"
keywords = ["ChannelExplorer"]
packages = [
    { include = "channelexplorer" },
]

[tool.poetry.dependencies]
python = "^3.10, <3.11"
numpy = "^1.26.3"
fastapi = "^0.108.0"
scikit-learn = "^1.3.2"
uvicorn = "^0.25.0"
tqdm = "^4.66.1"
pillow = "^10.2.0"
redis = "^5.0.1"
torch = {version = "^2.1.2", optional = true}
torchvision = {version = "^0.16.2", optional = true}
nptyping = "^2.5.0"
beartype = "^0.16.4"
tensorflow-datasets = {version = "^4.9.4", optional = true}
keract = {version = "^4.5.1", optional = true}
matplotlib = "^3.8.2"
grandalf = "^0.8"
nltk = "^3.8.1"
pydot = "^2.0.0"
umap-learn = "^0.5.5"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
tensorflow = "2.15"
transformers = "^4.44.2"
datasets = "^3.0.0"
pyvis = "^0.3.2"
diffusers = "^0.30.3"

[tool.poetry.extras]
tensorflow = ["tensorflow", "tensorflow-datasets", "keract"]
pytorch = ["torch", "torchvision"]

[tool.poetry.dev-dependencies]
ipykernel = "^6.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
